fix: algumas correções simples.

+++ backend/src/main/java/com/conryfinance/service/CardService.java
@@ -1,5 +1,6 @@
 package com.conryfinance.service;
 
+import com.conryfinance.model.enums.CardType;
 import com.conryfinance.model.monthlybudget.Card;
 import com.conryfinance.model.monthlybudget.MonthlyBudget;
 import com.conryfinance.repository.CardRepository;
@@ -21,7 +22,6 @@ public class CardService extends BaseService<Card, CardRepository> {
         Card defaultCardInstance = montlyBudgetComponentService.createDefaultCardInstance();
         MonthlyBudget monthlyBudget = monthlyBudgetService.findById(montlyBudgetId);
         defaultCardInstance.setMonthlyBudget(monthlyBudget);
-        repository.persist(defaultCardInstance);
         return defaultCardInstance;
     }
 
@@ -29,16 +29,19 @@ public class CardService extends BaseService<Card, CardRepository> {
     @Override
     public void deleteById(Long id) {
         Card card = findById(id);
+
+        if (card.getCardType().equals(CardType.TOTAL_AMOUNT_SPENT) || card.getCardType().equals(CardType.TOTAL_AVAILABLE)) {
+            throw new IllegalArgumentException("Card with this type cannot be deleted");
+        }
+
         card.setMonthlyBudget(null);
-        repository.persistAndFlush(card);
         super.deleteById(id);
     }
 
     @Transactional
-    public Card updateDescription(Long cardId, String newDescription) {
+    public Card updateDescription(final Long cardId, final String newDescription) {
         Card card = findById(cardId);
         card.setDescription(newDescription);
-        repository.persist(card);
         return card;
     }
 }
